pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "test-rag-2"
version = "0.1.0"
description = ""
authors = [
    {name = "Mahee Gamage",email = "mahee.gamage@gmail.com"}
]
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-core (>=1.2.14,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langgraph (>=1.0.9,<2.0.0)",
    "chromadb (>=1.5.1,<2.0.0)",
    "pypdf (>=6.7.1,<7.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "langchain-chroma (>=0.2.0,<0.3.0)",
    "tqdm (>=4.0.0,<5.0.0)",
    "streamlit (>=1.30.0)",
    "langsmith (>=0.7.6,<0.8.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langchain-openai (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
